refactor(api): split auth.service.ts into services/auth/ (api-huu3)
