+++ apps/api/src/__tests__/routes/account/change-password.test.ts
@@ -20,7 +20,7 @@ vi.mock("../../../services/account.service.js", () => ({
   },
 }));
 
-vi.mock("../../../services/auth.service.js", () => ({
+vi.mock("../../../services/auth/register.js", () => ({
   ValidationError: class ValidationError extends Error {
     override readonly name = "ValidationError" as const;
   },
@@ -37,7 +37,7 @@ vi.mock("../../../middleware/auth.js", () => mockAccountOnlyAuthFactory());
 
 const { createAuditWriter } = await import("../../../lib/audit-writer.js");
 const { changePassword, ConcurrencyError } = await import("../../../services/account.service.js");
-const { ValidationError } = await import("../../../services/auth.service.js");
+const { ValidationError } = await import("../../../services/auth/register.js");
 const { accountRoutes } = await import("../../../routes/account/index.js");
 
 // ── Helpers ──────────────────────────────────────────────────────
