+++ apps/api/src/__tests__/trpc/routers/auth.integration.test.ts
@@ -33,7 +33,8 @@ import { requireSession } from "../../../lib/auth-context.js";
 import { hashEmail } from "../../../lib/email-hash.js";
 import { checkRateLimit } from "../../../middleware/rate-limit.js";
 import { _resetAccountLoginStoreForTesting } from "../../../middleware/stores/account-login-store.js";
-import { listSessions, loginAccount } from "../../../services/auth.service.js";
+import { loginAccount } from "../../../services/auth/login.js";
+import { listSessions } from "../../../services/auth/sessions.js";
 import { dispatchWebhookEvent } from "../../../services/webhook-dispatcher.js";
 import { authRouter } from "../../../trpc/routers/auth.js";
 import { noopAudit, registerTestAccount } from "../../helpers/integration-setup.js";
